+++ installer.py
@@ -593,7 +593,7 @@ def check_diffusers():
     t_start = time.time()
     if args.skip_all or args.skip_git:
         return
-    sha = '56d438727036b0918b30bbe3110c5fe1634ed19d' # diffusers commit hash
+    sha = 'c052791b5fe29ce8a308bf63dda97aa205b729be' # diffusers commit hash
     pkg = pkg_resources.working_set.by_key.get('diffusers', None)
     minor = int(pkg.version.split('.')[1] if pkg is not None else -1)
     cur = opts.get('diffusers_version', '') if minor > -1 else ''